+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [3.1.8] - 2023.09.22
 ### Added
-- Generic extension point for Extension Ecosystem
+- Capability to initialize `TatumSDK` with various extensions from Extension Ecosystem using the init() method and `TatumConfig.configureExtensions`.
+- Encapsulation of `typedi` functionalities within `TatumSdkContainer`, ensuring that there are no enforced dependencies on the extensions.
+- `TatumSdkExtension` abstract class that allows extension creators to utilize services or other extensions from withing the `TatumSdkContainer`.
+- `async init()` method call for each registered extension
+- `destroy()` method call for each registered extension
 
 ## [3.1.7] - 2023.09.22
 ### Fixed